+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Optimize worker storage in pool.
 - Optimize worker alive status check.
+- Optimize `LESS_RECENTLY_USED` worker choice strategy.
 
 ### Fixed